examples/pc_serial_bridge/requirements.txt
# ESP32 Bridge Client Requirements
# Install with: pip install -r requirements.txt

pyserial>=3.5